update(telegrafs/components): rename import paths
