+++ ui/src/style/chronograf.scss
@@ -57,8 +57,8 @@
 @import 'src/shared/components/BoxTooltip.scss';
 @import 'src/shared/components/dropdown_auto_refresh/AutoRefreshDropdown.scss';
 @import 'src/organizations/components/Retention.scss';
-@import 'src/organizations/components/TelegrafConfigOverlay.scss';
-@import 'src/organizations/components/TelegrafExplainer.scss';
+@import 'src/telegrafs/components/TelegrafConfigOverlay.scss';
+@import 'src/telegrafs/components/TelegrafExplainer.scss';
 @import 'src/variables/components/CreateVariableOverlay.scss';
 @import 'src/tasks/components/TaskForm.scss';
 @import 'src/tasks/components/TasksPage.scss';
