+++ ui/src/telegrafs/components/Collectors.tsx
@@ -6,10 +6,10 @@ import {connect} from 'react-redux'
 // Components
 import {Input, Button, EmptyState, Grid} from '@influxdata/clockface'
 import {Tabs} from 'src/clockface'
-import CollectorList from 'src/organizations/components/CollectorList'
-import TelegrafExplainer from 'src/organizations/components/TelegrafExplainer'
-import TelegrafInstructionsOverlay from 'src/organizations/components/TelegrafInstructionsOverlay'
-import TelegrafConfigOverlay from 'src/organizations/components/TelegrafConfigOverlay'
+import CollectorList from 'src/telegrafs/components/CollectorList'
+import TelegrafExplainer from 'src/telegrafs/components/TelegrafExplainer'
+import TelegrafInstructionsOverlay from 'src/telegrafs/components/TelegrafInstructionsOverlay'
+import TelegrafConfigOverlay from 'src/telegrafs/components/TelegrafConfigOverlay'
 import CollectorsWizard from 'src/dataLoaders/components/collectorsWizard/CollectorsWizard'
 import FilterList from 'src/shared/components/Filter'
 import NoBucketsWarning from 'src/organizations/components/NoBucketsWarning'
